Warnings

+++ src/audio/CSoundMixer.h
@@ -68,7 +68,7 @@ public:
     MixerInfo *infoTable = 0;
     MixerInfo **sortSpace[2] = {0};
 
-    uint64_t baseTime = 0;
+    uint32_t baseTime = 0;
 
     int32_t frameTime = 0;
     int32_t frameStartTime = 0;
